requirements.txt
# requirements.txt 

# --- Core Frameworks ---

crewai>=0.35.0,<0.36.0
fastapi>=0.111.0,<0.112.0
uvicorn[standard]>=0.30.1,<0.31.0
streamlit>=1.36.0,<1.37.0
python-dotenv>=1.0.1,<2.0.0

# --- LLM & Embeddings (Gemini) ---

google-generativeai>=0.4.1,<0.5.0
# Use the latest 0.0.x version compatible with LangChain 0.1.x
langchain-google-genai>=0.0.11,<0.0.12

# --- RAG & Document Processing ---
langchain>=0.1.16,<0.1.17
langchain-core>=0.1.48,<0.1.49
langchain-community>=0.0.34,<0.0.35

# ChromaDB version compatible with LangChain 0.1.x integrations
chromadb>=0.4.24,<0.5.0
pypdf>=4.2.0,<4.3.0
sentence-transformers>=2.7.0,<2.8.0

# --- Utilities ---
requests>=2.32.3,<2.33.0